# organization-service/src/main/kotlin/emented/model/domain/Organization.kt
package emented.model.domain

import java.time.OffsetDateTime

data class Organization(
    val id: Long?,
    val name: String,
    val officialAddress: Address,
    val coordinates: Coordinates,
    val annualTurnover: Int?,
    val type: OrganizationType?,
    val employees: List<Employee>,
    val creationDate: OffsetDateTime?,
)